lATIHAN WEB 26/05/2025

+++ latihan/app/Http/Controllers/ProdiController.php
@@ -2,7 +2,9 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Fakultas;
 use App\Models\Prodi;
+use DB;
 use Illuminate\Http\Request;
 
 class ProdiController extends Controller
@@ -12,9 +14,11 @@ class ProdiController extends Controller
      */
     public function index()
     {
-        $data = Prodi::all();
-        return view("prodi.index",
-        compact('data'));
+        $listprodi = Prodi::all(); //select * from prodis;
+        //$listprodi = DB::table("prodis")->get();
+        return view("prodi.index", 
+            ['listprodi' => $listprodi]
+        );
     }
 
     /**
@@ -22,51 +26,63 @@ class ProdiController extends Controller
      */
     public function create()
     {
-        return view("prodi.create");
+        $fakutas = Fakultas::all();
+        return view("prodi.create", [
+            'fakultas' => $fakutas
+        ]);
+
     }
-
-    public function delete()
-    {
-        return view("prodi.delete");
-    }
-
-    public function detail()
-    {
-        return view("prodi.detail");
-    }
-
-    public function prodi() {
-        return view('materi.prodi');
-    }
-
-
 
     /**
      * Store a newly created resource in storage.
      */
     public function store(Request $request)
     {
-        $request->validate([
-        'nama_prodi' => 'required|string|max:255',
-    ]);
+        //Form Validation
+        $data = $request->validate([
+            'kode_prodi' => 'required|min:2|max:2',
+            'nama' => 'required|min:5|max:25',
+            'fakultas_id'=> 'required'
+        ]);
 
-    Prodi::create([
-        'nama_prodi' => $request->nama_prodi,
-    ]);
+        //$data = $request->all();
+        //cara 1
+        //response object dari created data
+        Prodi::create([
+            'kode_prodi' => $data['kode_prodi'],
+            'nama'          => $data['nama'],
+            'fakultas_id' => $data['fakultas_id']
+        ]);
+        
+        //cara 2
+        //response true atau false
+        // Prodi::insert([
+        //     'kode_prodi' => $data['kode_prodi'],
+        //     'nama' => $data['nama'],
+        // ]);
+        
+        //cara 3
+        // $newprodi = new Prodi();
+        // $newprodi->kode_prodi = $data['kode_prodi'];
+        // $newprodi->nama = $data['nama'];
+        // $newprodi->save();
 
-    return redirect()->route('prodi.index')->with('success', 'Prodi berhasil ditambahkan');
+        //arahkan/pindahkan ke halaman tujuan
+        return redirect("prodi")->with("status", "Program Studi berhasil disimpan!");
     }
 
     /**
      * Display the specified resource.
      */
-    public function show($id)
+    public function show(string $id)
     {
-      
-            $data = Prodi::find0rFail($id);
-            return view("prodi.show",
-            compact('data'));
-        
+        $prodi = Prodi::find($id);
+        if(!isset($prodi->id)){
+            return redirect("prodi")->with("failed", "Program Studi tidak ditemukan!");
+        }
+        return view("prodi.detail", [
+            'prodi' => $prodi
+        ]);
     }
 
     /**
@@ -74,8 +90,17 @@ class ProdiController extends Controller
      */
     public function edit(string $id)
     {
-        $data = Prodi::findOrFail($id);
-        return view('prodi.edit', compact('data'));
+        //Ambil data berdasarkan id
+        $prodi = Prodi::find($id); 
+        if(!isset($prodi->id)){
+            return redirect("prodi")->with("failed", "Program Studi tidak ditemukan!");
+        }
+
+        //select * from prodis where id = $id
+        //kirma data ke view
+        return view("prodi.edit", [
+            'prodi' => $prodi
+        ]);
     }
 
     /**
@@ -83,16 +108,19 @@ class ProdiController extends Controller
      */
     public function update(Request $request, string $id)
     {
-     $request->validate([
-            'nama_prodi' => 'required|string|max:255'
+        //Form Validation
+        $data = $request->validate([
+            //'kode_prodi' => 'required|min:2|max:2',
+            'nama' => 'required|min:5|max:25'
         ]);
+        //update data
+        $prodi = Prodi::find($id);
+        //$prodi->kode_prodi = $data['kode_prodi'];
+        $prodi->nama = $data['nama'];
+        $prodi->save();
 
-        $prodi = Prodi::findOrFail($id);
-        $prodi->update([
-            'nama_prodi' => $request->nama_prodi
-        ]);
-
-        return redirect()->route('prodi.index')->with('success', 'Data berhasil diupdate.');
+        return redirect("prodi")
+            ->with("status", "Program Studi berhasil diupdate!");
     }
 
     /**
@@ -100,11 +128,16 @@ class ProdiController extends Controller
      */
     public function destroy(string $id)
     {
-        $prodi = Prodi::find0rFail($id);
-        $prodi->delete();
+        $prodi = Prodi::find($id);
 
-        return
-        redirect()->route('prodi.index')->with('success', 'prodi berhail dihapus');
+        if(isset($prodi->id)){
+            $prodi->delete();
+            return redirect("prodi")->with("status", "Program Studi berhasil dihapus!");
+        }
+
+        return redirect("prodi")->with("failed", "Program Studi gagal dihapus!");
+        // $delete = DB::table("prodis")
+        //     ->where("id", $id)
+        //     ->delete();
     }
-
 }
